requirements.txt
Babel==2.7.0
# billiard==3.6.1.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-crispy-forms==1.8.0
django-npm==1.0.0
django-phonenumber-field==3.0.1
et-xmlfile==1.0.1
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
more-itertools==7.2.0
npm==0.1.1
oauthlib==3.1.0
openpyxl==2.5.1
optional-django==0.1.0
phonenumbers==8.10.22
psycopg2-binary==2.8.4
python3-openid==3.1.0
# pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.6
vine==1.3.0
XlsxWriter==1.2.5
zipp==0.6.0
